[issue-359] Move relationship_filters to model package
